+++ lib/recommendation/ui/pages/information_recommendation_page.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:myecl/recommendation/providers/recommendation_provider.dart';
-import 'package:myecl/recommendation/ui/pages/main_page/recommendation_card.dart';
-import 'package:myecl/recommendation/ui/pages/main_page/recommendation_card_layout.dart';
+import 'package:myecl/recommendation/ui/widgets/recommendation_card.dart';
+import 'package:myecl/recommendation/ui/widgets/recommendation_card_layout.dart';
 import 'package:myecl/recommendation/ui/widgets/recommendation_template.dart';
 
 class InformationRecommendationPage extends HookConsumerWidget {
