+++ lib/recommendation/ui/pages/main_page.dart
@@ -6,8 +6,8 @@ import 'package:myecl/recommendation/providers/is_recommendation_admin_provider.
 import 'package:myecl/recommendation/providers/recommendation_list_provider.dart';
 import 'package:myecl/recommendation/providers/recommendation_provider.dart';
 import 'package:myecl/recommendation/router.dart';
-import 'package:myecl/recommendation/ui/pages/main_page/recommendation_card.dart';
-import 'package:myecl/recommendation/ui/pages/main_page/recommendation_card_layout.dart';
+import 'package:myecl/recommendation/ui/widgets/recommendation_card.dart';
+import 'package:myecl/recommendation/ui/widgets/recommendation_card_layout.dart';
 import 'package:myecl/recommendation/ui/widgets/recommendation_template.dart';
 import 'package:myecl/tools/ui/builders/async_child.dart';
 import 'package:myecl/tools/ui/layouts/refresher.dart';
